fixed possible bug on calc_splizvd.py line 257
I am running this pipeline from the sicilian_called_splice_juncs.tsv file which the sicilian pipeline delivers (https://github.com/salzman-lab/SICILIAN). When I do not run it with the "light" mode, and since I have the column "missing_domains" present in the input table, it raises an error on line 261 as there is no column named posStart_group or posEnd_group, they have not been created before in the code of this script. However, there are two columns called posA_group and posB_group created upstream in this script using the start and end junction positions. I do not know if I am right with this change but I think that is possible that is pointing to this columns that already exist.